+++ examples/combo_box_table/combo_box_column_width/index.jsx
@@ -29,7 +29,7 @@ function index({ breed }) {
 function renderValue({ children }, { search, column: { name } }) {
   if (name === 'breed') {
     return (
-      <TokenHighlight search={search} label={children} />
+      <TokenHighlight search={search} value={children} />
     );
   }
   return children;