+++ src/components/highlighters/delimited_highlight.test.jsx
@@ -4,12 +4,12 @@ import { render } from '@testing-library/react';
 import { DelimitedHighlight } from './delimited_highlight';
 
 it('highlights empty string', () => {
-  const { container } = render(<DelimitedHighlight label="" start="<em>" end="</em>" />);
+  const { container } = render(<DelimitedHighlight value="" start="<em>" end="</em>" />);
   expect(container).toMatchInlineSnapshot(`<div />`);
 });
 
 it('highlights with no match', () => {
-  const { container } = render(<DelimitedHighlight label="foo" start="<em>" end="</em>" />);
+  const { container } = render(<DelimitedHighlight value="foo" start="<em>" end="</em>" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   foo
@@ -18,7 +18,7 @@ it('highlights with no match', () => {
 });
 
 it('highlights with a match', () => {
-  const { container } = render(<DelimitedHighlight label="foo <em>bar</em>" start="<em>" end="</em>" />);
+  const { container } = render(<DelimitedHighlight value="foo <em>bar</em>" start="<em>" end="</em>" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   <span
@@ -39,7 +39,7 @@ it('highlights with a match', () => {
 });
 
 it('inverses a highlight', () => {
-  const { container } = render(<DelimitedHighlight label="foo <em>bar</em>" start="<em>" end="</em>" inverse />);
+  const { container } = render(<DelimitedHighlight value="foo <em>bar</em>" start="<em>" end="</em>" inverse />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   <span