+++ src/components/highlighters/index.js
@@ -1,3 +1,4 @@
+export { Highlight } from './highlight';
 export { DelimitedHighlight } from './delimited_highlight';
 export { PrefixHighlight } from './prefix_highlight';
 export { SubstringHighlight } from './substring_highlight';