+++ src/components/highlighters/prefix_highlight.test.jsx
@@ -4,12 +4,12 @@ import { render } from '@testing-library/react';
 import { PrefixHighlight } from './prefix_highlight';
 
 it('highlights empty string', () => {
-  const { container } = render(<PrefixHighlight label="" search="" />);
+  const { container } = render(<PrefixHighlight value="" search="" />);
   expect(container).toMatchInlineSnapshot(`<div />`);
 });
 
 it('highlights with an empty search', () => {
-  const { container } = render(<PrefixHighlight label="foo" search="" />);
+  const { container } = render(<PrefixHighlight value="foo" search="" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   foo
@@ -18,7 +18,7 @@ it('highlights with an empty search', () => {
 });
 
 it('highlights with no match', () => {
-  const { container } = render(<PrefixHighlight label="foobar foo" search="bar" />);
+  const { container } = render(<PrefixHighlight value="foobar foo" search="bar" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   foobar foo
@@ -27,7 +27,7 @@ it('highlights with no match', () => {
 });
 
 it('highlights with a match', () => {
-  const { container } = render(<PrefixHighlight label="foobar foo" search="foo" />);
+  const { container } = render(<PrefixHighlight value="foobar foo" search="foo" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   <span
@@ -48,7 +48,7 @@ it('highlights with a match', () => {
 });
 
 it('inverses the a highlight', () => {
-  const { container } = render(<PrefixHighlight label="foobar foo" search="foo" inverse />);
+  const { container } = render(<PrefixHighlight value="foobar foo" search="foo" inverse />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   <span