+++ src/components/highlighters/token_highlight.test.jsx
@@ -4,12 +4,12 @@ import { render } from '@testing-library/react';
 import { TokenHighlight } from './token_highlight';
 
 it('highlights empty string', () => {
-  const { container } = render(<TokenHighlight label="" search="" />);
+  const { container } = render(<TokenHighlight value="" search="" />);
   expect(container).toMatchInlineSnapshot(`<div />`);
 });
 
 it('highlights with an empty search', () => {
-  const { container } = render(<TokenHighlight label="foo" search="" />);
+  const { container } = render(<TokenHighlight value="foo" search="" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   foo
@@ -18,7 +18,7 @@ it('highlights with an empty search', () => {
 });
 
 it('highlights with no match', () => {
-  const { container } = render(<TokenHighlight label="foo foobar" search="bar" />);
+  const { container } = render(<TokenHighlight value="foo foobar" search="bar" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   foo foobar
@@ -27,7 +27,7 @@ it('highlights with no match', () => {
 });
 
 it('highlights with a match', () => {
-  const { container } = render(<TokenHighlight label="foo foobar" search="foo" />);
+  const { container } = render(<TokenHighlight value="foo foobar" search="foo" />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   <span
@@ -52,7 +52,7 @@ it('highlights with a match', () => {
 });
 
 it('inverses the a highlight', () => {
-  const { container } = render(<TokenHighlight label="foo foobar" search="foo" inverse />);
+  const { container } = render(<TokenHighlight value="foo foobar" search="foo" inverse />);
   expect(container).toMatchInlineSnapshot(`
 <div>
   <span